[grid] in gg_tt.mad/bin/internal/madevent_interface.py, avoid deleting GpuAbstraction.h and GpuRuntime.h #947

+++ epochX/cudacpp/gg_tt.mad/bin/internal/madevent_interface.py
@@ -3967,6 +3967,8 @@ Beware that this can be dangerous for local multicore runs.""")
         for P in Pdir: 
             allG = misc.glob('G*', path=P)
             for G in allG:
+                if G.endswith('GpuAbstraction.h'): continue
+                if G.endswith('GpuRuntime.h'): continue
                 if pjoin(P, G) not in Gdir:
                     logger.debug('removing %s', pjoin(P,G))
                     shutil.rmtree(pjoin(P,G))
